requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bleach==6.2.0
bokeh==3.4.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.0
distro==1.9.0
exceptiongroup==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
numpy==2.0.2
openai==1.63.2
packaging==24.2
pandas==2.2.3
panel==1.4.5
param==2.2.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
xyzservices==2025.1.0
zipp==3.21.0
